requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.1.1
cryptography==41.0.7
cycler==0.12.1
exceptiongroup==1.2.0
filelock==3.13.1
fonttools==4.45.1
fsspec==2023.10.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.2
kaitaistruct==0.10
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.4
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
pyasn1==0.5.1
pycparser==2.21
pyOpenSSL==23.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
selenium==4.15.2
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sympy==1.12
torch==2.1.1
torchvision==0.16.1
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wsproto==1.2.0
zipp==3.17.0
zstandard==0.22.0
